Refactored code  - Moved getInteger to Math utility file  - Capatilized utility files
