+++ 100-js-algo/centuryFromYear.js
@@ -1,4 +1,4 @@
-const { getInteger } = require("./utils/utils");
+const { getInteger } = require("./utils/math");
 function centuryFromYear(year) {
     if (year < 1) {
         return "Please provide proper year";
@@ -7,8 +7,8 @@ function centuryFromYear(year) {
     if (year % 100 === 0) {
         return century;
     }
-    
-return century + 1;
+
+    return century + 1;
 }
 
 module.exports = { centuryFromYear };
