+++ 100-js-algo/circleOfNumbers.js
@@ -5,7 +5,7 @@ function circleOfNumbers(numberOfPlayers, playerNumber) {
     if (playerNumber > numberOfPlayers) {
         return "Player does not exist";
     }
-    const { getInteger } = require("./utils/utils");
+    const { getInteger } = require("./utils/math");
     const split = getInteger(numberOfPlayers / 2);
 
     return playerNumber + split;
